Ajout uml + maven
