+++ .gitignore
@@ -1,9 +1,13 @@
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
+
 .settings
 .vscode
 .classpath
 .project
 target
-.gradle
-build
 gradle/
 solution
