+++ build.gradle
@@ -1,0 +1,51 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * This generated file contains a sample Java Library project to get you started.
+ * For more details take a look at the Java Libraries chapter in the Gradle
+ * User Manual available at https://docs.gradle.org/6.0/userguide/java_library_plugin.html
+ */
+
+plugins {
+    // Apply the java-library plugin to add support for Java Library
+    id 'java'
+    id 'application'
+    
+    // PLant Uml plugin
+    id "com.github.roroche.plantuml" version "1.0.2"
+}
+
+repositories {
+    // Use jcenter for resolving dependencies.
+    // You can declare any Maven/Ivy/file repository here.
+    jcenter()
+}
+
+dependencies {
+    // This dependency is used internally, and not exposed to consumers on their own compile classpath.
+    implementation 'com.google.guava:guava:28.2-jre'
+
+    // Use JUnit test framework
+    testImplementation 'junit:junit:4.12'
+}
+
+application {
+    // Define the main class for the application.
+    mainClassName = 'tp3.Appli'
+}
+
+jar {
+    destinationDirectory.set(file("target/"))
+    baseName 'pizzeria'
+
+    manifest { 
+        attributes 'Main-Class': 'tp3.Appli' 
+    } 
+
+}
+
+
+classDiagram { // (1)
+    packageName = "tp3" // (2)
+    outputFile = project.file('docs/TP3.plantuml') // (3)
+}
