+++ src/main/java/tp3/Appli.java
@@ -1,3 +1,4 @@
+package tp3;
 
 public class Appli {
