+++ src/main/java/tp3/Pizzeria.java
@@ -1,3 +1,4 @@
+package tp3;
 /**
  * @author bruel (from O'Reilly Head-First series)
  * @depend - * - Pizza
