+++ src/main/java/tp3/PizzeriaBrest.java
@@ -1,17 +1,18 @@
+package tp3;
 
-public class PizzeriaStrasbourg extends Pizzeria {
-	
+public class PizzeriaBrest extends Pizzeria {
+
 	@Override
 	protected
 	Pizza creerPizza(String type) {
 		
 		if (type.equals("fromage")) {
-			return new PizzaFromageStyleStrasbourg();
+			return new PizzaFromageStyleBrest();
 		} else if (type.equals("grecque")) {
 			return new PizzaGrecque();
 		} else {
 			return new PizzaPoivrons();
 		}
 	}
-	
+
 }
