+++ src/main/java/tp3/SimpleFabriqueDePizzas.java
@@ -1,3 +1,4 @@
+package tp3;
 
 public abstract class SimpleFabriqueDePizzas {
